+++ axion/specification/__init__.py
@@ -1,0 +1,23 @@
+from axion.specification import loader
+from axion.specification.model import (
+    OASContent,
+    OASMediaType,
+    OASOperation,
+    OASParameter,
+    OASServer,
+    OASSpecification,
+    OASType,
+)
+
+load = loader.load_spec
+
+__all__ = [
+    'load',
+    'OASSpecification',
+    'OASServer',
+    'OASType',
+    'OASOperation',
+    'OASContent',
+    'OASParameter',
+    'OASMediaType',
+]
